+++ src/test/java/lotto/LottoDispenserTest.java
@@ -2,6 +2,10 @@ package lotto;
 
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.domain.LottoDispenser;
+import lotto.domain.LottoTickets;
+import lotto.domain.NumberSelectStrategy;
+import lotto.domain.RandomNumberSelectStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -15,7 +19,7 @@ public class LottoDispenserTest {
     @MethodSource("getIssueLottoTicketByPriceData")
     public void issue_lotto_ticket_by_price(int price, int number) {
         LottoDispenser lottoDispenser = new LottoDispenser(new RandomNumberSelectStrategy());
-        Assertions.assertThat(lottoDispenser.getLottoTicket(price).size()).isEqualTo(number);
+        Assertions.assertThat(lottoDispenser.getLottoTicket(price).getCount()).isEqualTo(number);
     }
 
     private static Stream<Arguments> getIssueLottoTicketByPriceData() {
@@ -29,27 +33,23 @@ public class LottoDispenserTest {
     @DisplayName("숫자 생성 전략을 통한 숫자 생성")
     @ParameterizedTest
     @MethodSource("getIssueLottoTicketByNumberSelectStrategyData")
-    public void issue_lotto_ticket_by_price(List<List<Integer>> randomNumbers, int price, List<String> expected) {
+    public void issue_lotto_ticket_by_price(List<List<Integer>> randomNumbers, int price, String expected) {
         LottoDispenser lottoDispenser = new LottoDispenser(createNumberSelectStrategy(randomNumbers));
-        List<LottoTicket> lottoTickets = lottoDispenser.getLottoTicket(price);
-        for (int i = 0; i < lottoTickets.size(); i++) {
-            Assertions.assertThat(lottoTickets.get(i).getString()).isEqualTo(expected.get(i));
-        }
+        LottoTickets lottoTickets = lottoDispenser.getLottoTicket(price);
+        Assertions.assertThat(lottoTickets.getLottoNumbersString()).isEqualTo(expected);
     }
 
     private static Stream<Arguments> getIssueLottoTicketByNumberSelectStrategyData() {
         return Stream.of(
-                Arguments.of(List.of(List.of(1, 2, 3, 4, 5, 6)), 1000, List.of("[1, 2, 3, 4, 5, 6]")),
-                Arguments.of(
-                        List.of(
+                Arguments.of(List.of(
+                                List.of(1, 2, 3, 4, 5, 6)
+                        ),
+                        1000, "[1, 2, 3, 4, 5, 6]"),
+                Arguments.of(List.of(
                                 List.of(1, 2, 3, 4, 5, 6),
                                 List.of(5, 6, 20, 23, 40, 41)
                         ),
-                        2500,
-                        List.of(
-                                "[1, 2, 3, 4, 5, 6]",
-                                "[5, 6, 20, 23, 40, 41]"
-                        )
+                        2500, "[1, 2, 3, 4, 5, 6]\n[5, 6, 20, 23, 40, 41]"
                 )
         );
     }
