+++ src/test/java/lotto/LottoGameTest.java
@@ -1,6 +1,8 @@
 package lotto;
 
 import java.util.stream.Stream;
+import lotto.domain.LottoGame;
+import lotto.domain.RandomNumberSelectStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
