+++ src/test/java/lotto/LottoNumbersTest.java
@@ -3,6 +3,8 @@ package lotto;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoNumbers;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
