+++ src/test/java/lotto/LottoTicketsTest.java
@@ -1,6 +1,10 @@
 package lotto;
 
 import java.util.Arrays;
+import lotto.domain.LottoResult;
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTickets;
+import lotto.domain.LottoWinningNumbers;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
